net_collector/requirements.txt
netmiko>=4.0.0
PyYAML>=6.0
paramiko>=4.0.0
